Sush/validation checks

+++ controllers/api/reading-list-routes.js
@@ -2,6 +2,8 @@ const router = require('express').Router();
 const { Reading_List, Book_Reading_List, Book } = require('../../models');
 const withAuth = require('../../utils/auth');
 
+// path: /api/reading-list
+
 // get all book reading lists (for testing purposes)
 router.get('/', async (req, res) => {
   try {
@@ -12,8 +14,6 @@ router.get('/', async (req, res) => {
     res.status(500).json(error);
   }
 });
-
-// path: /api/reading-list
 
 // WHEN USER ADDS A BOOK TO THEIR READING LIST
 router.post('/add', withAuth, async (req, res) => {
@@ -35,30 +35,46 @@ router.post('/add', withAuth, async (req, res) => {
       }
     });
 
-    // convert Sequelize object into plain object
-    // readingListData.get({ plain: true });
-
-    // grab the reading list ID out of the plain obj
+    // grab the reading list ID out of the returned data
     const readingListId = readingListData.id;
 
     // console.log('\n---RL ROUTE: GET READING LIST');
     // console.log(userId, readingListId, bookId);
     // console.log(book);
 
+    // first see if the book exists on the user's reading list
+    const isBookOnReadingList = await Book_Reading_List.findOne({
+      where: {
+        book_id: bookId,
+        reading_list_id: readingListId
+      }
+    });
+
+    // console.log('\n---RL ROUTES: DOES BOOKLIST EXIST');
+    // console.log(isBookOnReadingList);
+
+    // if a bookList does NOT exist (ie. a book is NOT on a user's reading list), then go ahead and create a new match
+    if (!isBookOnReadingList) {
+      await Book_Reading_List.create({
+        book_id: bookId,
+        reading_list_id: readingListId
+      });
+
+      // console.log(bookReadingList);
+
+      // send details of the book that is to be added to the reading list
+      res.status(200).json({
+        book
+      });
+    } else {
+      res.status(400).json({
+        message: 'Book has already been added to reading list'
+      });
+    }
+
     // make the attachment to the book-reading-list join table
-    await Book_Reading_List.create({
-      book_id: bookId,
-      reading_list_id: readingListId
-    });
-
-    // console.log(bookReadingList);
-
-    // send details of book that is to be added to reading list
-    res.status(200).json({
-      book
-    });
   } catch (error) {
-    console.log('\n---RL ROUTE: POST TO RL');
+    console.log('\n---RL ROUTE: POST TO RL ERR');
     console.log(error);
     res.status(500).json(error);
   }
@@ -70,34 +86,41 @@ router.delete('/remove', withAuth, async (req, res) => {
     const { bookId } = req.body;
     const userId = req.session.user_id;
 
-    // first find the correct book-reading-list ID using bookId and userId
-    const bookListData = await Book_Reading_List.findOne({
+    // first see if the book is already on the user's reading list
+    const isBookOnReadingList = await Book_Reading_List.findOne({
       where: {
         book_id: bookId,
         reading_list_id: userId
       }
     });
 
-    // store the id of the booklist to be removed in a new var
-    const bookListToBeRemoved = bookListData.id;
+    // console.log('\n---RL DELETE: IS BOOK ON READING LIST');
+    // console.log(isBookOnReadingList);
 
-    // console.log('\n---RL ROUTES: DELETE, FIND BOOK READING LIST');
-    // console.log(bookListData);
-    // console.log(bookListToBeRemoved);
+    // if the book IS on the user's reading list:
+    if (isBookOnReadingList) {
+      // grab the id of the booklist that we want to destroy
+      const bookToBeRemoved = isBookOnReadingList.id;
 
-    // remove the booklist column, which removes the book off the user's reading list
-    await Book_Reading_List.destroy({
-      where: {
-        id: bookListToBeRemoved.id
-      }
-    });
+      // destroy the link between book and reading list
+      await Book_Reading_List.destroy({
+        where: {
+          id: bookToBeRemoved
+        }
+      });
 
-    res.status(200).json({
-      bookListToBeRemoved,
-      message: 'Successfully removed book from reading list'
-    });
+      // send the data of the destroyed book and reading list link, and also a back-end message that it was successful
+      res.status(200).json({
+        bookToBeRemoved,
+        message: 'Successfully removed book from reading list'
+      });
+    } else {
+      res.status(400).json({
+        message: 'Book was never added to the reading list'
+      });
+    }
   } catch (error) {
-    console.log('\n---RL ROUTE: DELETE FROM RL');
+    console.log('\n---RL ROUTE: DELETE FROM RL ERR');
     console.log(error);
     res.status(500).json(error);
   }
